refactor(screener): move price-history/OHLCV shaping helpers to swing_screener.data.price_history
